General Code Improvements
Added lor - which is "lvalue_or_rvalue" taken from an stackoverflow answer by Matt Whitlock - to prevent unnecessary copies. 
Updated the README and removed the TODO-List also fixed a typo.
Also adjusted the helper methods as I was a little stupid and thought they should take a reference and not the extensions, so I changed that.